requirements.txt
pyaudio==0.2.13
requests==2.28.2
flask==2.2.3
flask-socketio==5.3.3
python-socketio==5.8.0
boto3==1.26.132
numpy==1.24.3 